consolidate radio and pinger into coms_launch
